backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-rest-framework-mongoengine==3.4.1
mongoengine==0.27.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
pymongo[srv]==4.6.0
Pillow==10.1.0
drf-spectacular==0.26.5
django-ratelimit==4.1.0
django-filter==23.5
whitenoise==6.6.0
